requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
numpy==1.26.3
pandas==2.1.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
